Start a benchmark suite (some code movement).
